+++ example/main.dart
@@ -57,6 +57,7 @@ class _MyHomePageState extends State<MyHomePage> {
       ListTile(leading: _buildAction4Color(), title: _buildColor()),
       ListTile(leading: _buildAction4HueSlider(), title: _buildHueSlider()),
       ListTile(leading: _buildAction4Ring(), title: _buildRing()),
+      ListTile(leading: _buildAction4Wheel(), title: _buildWheel()),
       ListTile(leading: _buildAction4Grid(), title: _buildGrid()),
       ListTile(leading: _buildAction4RGBSlider(), title: _buildRGBSlider()),
     ],
@@ -105,6 +106,12 @@ class _MyHomePageState extends State<MyHomePage> {
     tooltip: 'Open RGB chooser in dialog',
   );
 
+  Widget _buildAction4Wheel() => IconButton.outlined(
+    icon: Icon(Icons.lightbulb_circle_outlined),
+    onPressed: () => _showWheel(context, color),
+    tooltip: 'Open wheel chooser in dialog',
+  );
+
   Widget _buildColor() => Frame(
     label: 'Selected Color - #${color.toARGB32().toRadixString(16)}',
     child: Container(color: color, height: kMinInteractiveDimension),
@@ -137,25 +144,25 @@ class _MyHomePageState extends State<MyHomePage> {
 
   Widget _buildRing() => Frame(
     label: 'Ring Chooser',
-    child: Row(
-      children: [
-        SizedBox(
-          height: 200,
-          width: 200,
-          child: ColorusRing(
-            color: color,
-            onChanged: (col) => setState(() => color = col),
-          ),
-        ),
-        SizedBox(
-          height: 200,
-          width: 200,
-          child: ColorusWheel(
-            color: color,
-            onChanged: (col) => setState(() => color = col),
-          ),
-        ),
-      ],
+    child: SizedBox(
+      height: 200,
+      width: 200,
+      child: ColorusRing(
+        color: color,
+        onChanged: (col) => setState(() => color = col),
+      ),
+    ),
+  );
+
+  Widget _buildWheel() => Frame(
+    label: 'Wheel Chooser',
+    child: SizedBox(
+      height: 200,
+      width: 200,
+      child: ColorusWheelWithToggle(
+        color: color,
+        onChanged: (col) => setState(() => color = col),
+      ),
     ),
   );
 
@@ -198,13 +205,43 @@ class _MyHomePageState extends State<MyHomePage> {
   Future<Color?> _showRing(BuildContext context, Color clr) =>
       showAdaptiveDialog<Color?>(
         context: context,
-        builder: (BuildContext context) => AlertDialog(
-          title: Text('Colorus - Ring'),
-          content: ColorusRing(
-            color: clr,
-            onChanged: (col) => setState(() => color = col),
-          ),
-        ),
+        builder: (BuildContext context) {
+          Color dialogColor = clr;
+          return AlertDialog(
+            title: Text('Colorus - Ring'),
+            content: StatefulBuilder(
+              builder: (context, setState) => SizedBox(
+                width: 190,
+                height: 170,
+                child: ColorusRing(
+                  color: dialogColor,
+                  onChanged: (col) => setState(() => dialogColor = col),
+                ),
+              ),
+            ),
+          );
+        },
+      );
+
+  Future<Color?> _showWheel(BuildContext context, Color clr) =>
+      showAdaptiveDialog<Color?>(
+        context: context,
+        builder: (BuildContext context) {
+          Color dialogColor = clr;
+          return AlertDialog(
+            title: Text('Colorus - Ring'),
+            content: StatefulBuilder(
+              builder: (context, setState) => SizedBox(
+                width: 200,
+                height: 200,
+                child: ColorusWheelWithToggle(
+                  color: dialogColor,
+                  onChanged: (col) => setState(() => dialogColor = col),
+                ),
+              ),
+            ),
+          );
+        },
       );
 }