+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* Fixed selector jumps in ColorusRing
+* Added black/white toggle to ColorusWheel
+
 ## 1.0.0
 
 * Added ColorusWheel